Fixed all routes public and JWT auth, and one local auth

+++ routes/menuRoutes.js
@@ -1,21 +1,9 @@
 const express = require('express')
 const router = express.Router()
-
+const { jwtAuthMiddleware } = require('./../jwt')
 const MenuItem = require('../models/MenuItem')
 
-router.post('/', async (req, res) => {
-    try {
-        const data = req.body
-        const newMenuItem = new MenuItem(data)
-        const response = await newMenuItem.save()
-        console.log('Data saved successfully')
-        res.status(200).json(response)
-    } catch (err) {
-        console.log("Error in saving person", err)
-        res.status(500).json({ error: 'Internal server error' })
-    }
-})
-
+//[public] any body can view menu items
 router.get('/', async (req, res) => {
     try {
         const data = await MenuItem.find()
@@ -27,6 +15,7 @@ router.get('/', async (req, res) => {
     }
 })
 
+//[public] any body can filter menu items based on taste
 router.get('/:taste', async (req, res) => {
     try {
         const taste = req.params.taste
@@ -35,7 +24,7 @@ router.get('/:taste', async (req, res) => {
             console.log('response fetched');
             res.status(200).json(response)
         } else {
-            res.status(404).json({ error: 'Invalid work type' })
+            res.status(404).json({ error: 'Invalid taste type' })
         }
     } catch (error) {
         console.log("Error : ", error)
@@ -43,37 +32,67 @@ router.get('/:taste', async (req, res) => {
     }
 })
 
-router.put('/:id', async (req, res) => {
+//[jwt] only chef or manager can add menu item
+router.post('/', jwtAuthMiddleware, async (req, res) => {
     try {
-        const menuId = req.params.id;
-        const updatedMenuData = req.body
-        const response = await MenuItem.findByIdAndUpdate(menuId, updatedMenuData, {
-            new: true, //return the updated document
-            runValidators: true, //run mongoose validation
-        })
-        if (!response) {
-            //when person with id does not exist then response will be null
-            return res.status(404).json({ error: 'Menu not found' })
+        const userData = req.user
+        if (['chef', 'manager'].includes(userData.type)) {
+            const data = req.body
+            const newMenuItem = new MenuItem(data)
+            const response = await newMenuItem.save()
+            console.log('Data saved successfully')
+            res.status(200).json(response)
+        } else {
+            return res.status(401).json({ error: 'Not authorized' })
         }
-        console.log('data updated')
-        res.status(200).json(response)
+    } catch (err) {
+        console.log("Error in saving menu item", err)
+        res.status(500).json({ error: 'Internal server error' })
+    }
+})
+
+//[jwt] only manager / chef can edit menu items
+router.put('/:id', jwtAuthMiddleware, async (req, res) => {
+    try {
+        const userData = req.user
+        if (['chef', 'manager'].includes(userData.type)) {
+            const menuId = req.params.id;
+            const updatedMenuData = req.body
+            const response = await MenuItem.findByIdAndUpdate(menuId, updatedMenuData, {
+                new: true, //return the updated document
+                runValidators: true, //run mongoose validation
+            })
+            if (!response) {
+                //when menu with id does not exist then response will be null
+                return res.status(404).json({ error: 'Menu not found' })
+            }
+            console.log('data updated')
+            res.status(200).json(response)
+        } else {
+            return res.status(401).json({ error: 'Not authorized' })
+        }
     } catch (error) {
         console.log("Error : ", error)
         res.status(500).json({ error: 'Internal server error' })
     }
 })
 
-router.delete('/:id', async (req, res) => {
+//[jwt] only manager can delete menu items
+router.delete('/:id', jwtAuthMiddleware, async (req, res) => {
     try {
-        const menuId = req.params.id;
-
-        const response = await MenuItem.findByIdAndDelete(menuId)
-        if (!response) {
-            //when person with id does not exist then response will be null
-            return res.status(404).json({ error: 'MenuItem not found' })
+        const userData = req.user
+        if (userData.type == 'manager') {
+            const menuId = req.params.id;
+            const response = await MenuItem.findByIdAndDelete(menuId)
+            if (!response) {
+                //when menu with id does not exist then response will be null
+                return res.status(404).json({ error: 'MenuItem not found' })
+            }
+            console.log('data deleted')
+            res.status(200).json({ message: 'MenuItem deleted successfully' })
+        } else {
+            return res.status(401).json({ error: 'Not authorized' })
         }
-        console.log('data deleted')
-        res.status(200).json({ message: 'MenuItem deleted successfully' })
     } catch (error) {
         console.log("Error : ", error)
         res.status(500).json({ error: 'Internal server error' })
